server/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
Markdown==3.7
packaging==24.2
pillow==11.0.0
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
uritemplate==4.1.1
